requirements.txt
Flask
imutils
opencv-python-headless==4.5.5.64
Pillow
scikit-Image
gunicorn